Refactored root package, simplified CacheApi and extracted impl package
