+++ src/main/scala/play/api/cache/redis/Configuration.scala
@@ -4,8 +4,6 @@ import javax.inject.{Provider, Singleton}
 
 import scala.concurrent.duration.FiniteDuration
 
-import play.api.cache.redis.connector.RedisConnectorImpl
-
 /**
  * Configuration provider returns a connection settings to Redis server. The application can run in various
  * different environments and in many of them there might be very different connection settings. For example,
@@ -36,7 +34,7 @@ trait Configuration {
   /** When enabled security, this returns password for the AUTH command */
   def password: Option[ String ]
 
-  /** encapsulates connection settings into a single object used by [[RedisConnectorImpl]]. */
+  /** encapsulates connection settings into a single object used by [[play.api.cache.redis.connector.RedisConnectorImpl]]. */
   def connectionSettings( system: akka.actor.ActorSystem ) = ConnectionSettings(
     host, port, database, password
   )(
