+++ src/main/scala/play/api/cache/redis/connector/RedisConnectorModule.scala
@@ -1,5 +1,6 @@
 package play.api.cache.redis.connector
 
+import play.api.cache.redis.RedisConnector
 import play.api.{Configuration, Environment}
 import play.api.inject.{Binding, Module}
