+++ src/main/scala/play/api/cache/redis/impl/RedisCache.scala
@@ -1,11 +1,11 @@
-package play.api.cache.redis
+package play.api.cache.redis.impl
 
 import scala.concurrent._
 import scala.concurrent.duration.Duration
 import scala.language.{higherKinds, implicitConversions}
 import scala.reflect.ClassTag
 
-import play.api.cache.redis.connector.RedisConnector
+import play.api.cache.redis.{ConnectionSettings, InternalCacheApi, RedisConnector}
 
 /** <p>Implementation of plain API using redis-server cache and Brando connector implementation.</p> */
 class RedisCache[ Result[ _ ] ]( redis: RedisConnector, settings: ConnectionSettings )( implicit builder: Builders.ResultBuilder[ Result ] ) extends InternalCacheApi[ Result ] with Implicits {
