+++ src/main/scala/play/api/cache/redis/package.scala
@@ -3,4 +3,9 @@ package play.api.cache
 /**
   * @author Karel Cemus
   */
-package object redis extends AnyRef with Expiration
+package object redis extends AnyRef with Expiration {
+
+  type SynchronousResult[ A ] = A
+  type AsynchronousResult[ A ] = scala.concurrent.Future[ A ]
+
+}
